// server/src/utils/enums.ts
export enum Status {
    OPEN = "open",
    WAITLIST = "waitlist",
    FULL = "full"
}
  
export enum Semester {
    FALL = "fall",
    SUMMER = "summer",
    WINTER = "winter"
}

export enum Day {
    MONDAY = "monday",
    TUESDAY = "tuesday",
    WEDNESDAY = "wednesday",
    THURSDAY = "thursday",
    FRIDAY = "friday"
}